hotfix #113 그룹 이미지 처리방식 수정

Closes #113

See merge request s11-bigdata-recom-sub1/S11P21D208!139

+++ back/src/main/java/com/fastarm/back/team/service/TeamService.java
@@ -2,8 +2,6 @@ package com.fastarm.back.team.service;
 
 import com.fastarm.back.common.constants.S3Constants;
 import com.fastarm.back.common.service.S3Service;
-
-import com.fastarm.back.notification.entity.Notification;
 import com.fastarm.back.notification.entity.NotificationTeamInvite;
 import com.fastarm.back.notification.enums.Status;
 import com.fastarm.back.notification.enums.Type;
@@ -14,6 +12,7 @@ import com.fastarm.back.team.controller.dto.TeamInviteResponse;
 import com.fastarm.back.team.dto.*;
 import com.fastarm.back.team.entity.Team;
 import com.fastarm.back.team.entity.TeamMember;
+import com.fastarm.back.team.exception.TeamImageUploadException;
 import com.fastarm.back.team.exception.TeamMemberNotFoundException;
 import com.fastarm.back.team.repository.TeamRepository;
 import com.fastarm.back.team.repository.TeamMemberRepository;
@@ -29,7 +28,6 @@ import java.io.IOException;
 import java.net.URISyntaxException;
 import java.util.ArrayList;
 import java.util.List;
-
 
 @Service
 @RequiredArgsConstructor
@@ -53,7 +51,6 @@ public class TeamService {
 
     @Transactional(readOnly = true)
     public TeamDetailDto getTeamDetail(TeamDetailRequest dto) {
-
 
         Team team = teamRepository.findById(dto.getTeamId()).orElseThrow(TeamNotFoundException::new);
         Member member = memberRepository.findByLoginId(dto.getLoginId()).orElseThrow(MemberNotFoundException::new);
@@ -88,23 +85,22 @@ public class TeamService {
                 continue;
             }
 
-            if (notificationTeamInviteRepository.existsByTeamAndNotificationReceiverIdAndStatus(team,member.getId(), Status.WAIT)) {
+            if (notificationTeamInviteRepository.existsByTeamAndReceiverIdAndStatus(team,member.getId(), Status.WAIT)) {
                 previouslyInvited.add(receiverNickName);
                 continue;
             }
 
             successfulInvites.add(receiverNickName);
-            Notification notification = Notification.builder()
+
+            NotificationTeamInvite teamInvite = NotificationTeamInvite.builder()
                     .receiver(member)
                     .sender(sender)
                     .content(team.getName() + " 팀에 초대되었습니다.")
                     .type(Type.TEAM_INVITE)
-                    .build();
-            notificationRepository.save(notification);
-
-            NotificationTeamInvite teamInvite = NotificationTeamInvite.builder()
                     .team(team)
-                    .notification(notification)
+                    .isDeleted(false)
+                    .isRead(false)
+                    .status(Status.WAIT)
                     .build();
             notificationTeamInviteRepository.save(teamInvite);
         }
@@ -119,12 +115,10 @@ public class TeamService {
 
     @Transactional
     public Long createTeam(TeamAddDto dto) throws IOException {
-        String imagePath;
-        if (dto.getTeamImage().isEmpty()) {
-            imagePath = S3Constants.DEFAULT_IMAGE;
-        } else {
-            imagePath = s3Service.uploadFile(dto.getTeamImage());
-        }
+
+        if(dto.getTeamImage().isEmpty()) throw new TeamImageUploadException();
+        String imagePath = s3Service.uploadFile(dto.getTeamImage());
+
 
         Team team = dto.toEntity(imagePath);
         Team savedTeam = teamRepository.save(team);
@@ -151,6 +145,7 @@ public class TeamService {
         checkPermission(member,team);
 
         String imagPath;
+        if(dto.getTeamImage().isEmpty()) throw new TeamImageUploadException();
         if (dto.getTeamImage() != null) {
             s3Service.deleteFile(team.getTeamImage());
             imagPath = s3Service.uploadFile(dto.getTeamImage());
@@ -162,16 +157,14 @@ public class TeamService {
         teamRepository.save(team);
     }
 
-
     @Transactional
     public void checkPermission(Member member, Team team) {
         boolean isMember = teamMemberRepository.existsByTeamAndMember(team,member);
-
         if (!isMember) throw new TeamMemberNotFoundException();
     }
 
     @Transactional
-    public void withdrawTeam(TeamWithdrawDto dto){
+    public void withdrawTeam(TeamWithdrawDto dto) throws URISyntaxException {
         Member member = memberRepository.findByLoginId(dto.getLoginId()).orElseThrow(MemberNotFoundException::new);
         Team team = teamRepository.findById(dto.getTeamId()).orElseThrow(TeamNotFoundException :: new);
         TeamMember teamMember = teamMemberRepository.findByTeamIdAndMemberId(dto.getTeamId(),member.getId())
@@ -180,7 +173,9 @@ public class TeamService {
         teamMemberRepository.delete(teamMember);
 
         int remainingMembers = teamMemberRepository.countByTeamId(dto.getTeamId());
-        if (remainingMembers==0) teamRepository.delete(team);
+        if (remainingMembers==0){
+            s3Service.deleteFile(team.getTeamImage());
+            teamRepository.delete(team);
+        }
     }
-
 }
